+++ src/lib/vm.h
@@ -13,7 +13,7 @@ struct Worms {
   bool halted;
   bool full_trace;
   word jump_count = {.INT = 0};
-  word last_jmp = {.INT = static_cast<uint64_t>(-1)};
+  word last_jmp = {.INT = -1};
   word load_program(vector<Inst> prog);
   word exec();
   string debug(word trap);
